+++ service/controller/legacy/controller.go
@@ -15,12 +15,12 @@ import (
 
 type ControllerConfig struct {
 	BaseDomain       string
+	Provider         string
 	CreatePVC        bool
 	StorageSize      string
 	K8sClient        k8sclient.Interface
 	Logger           micrologger.Logger
 	PrometheusClient promclient.Interface
-	Provider         string
 }
 
 type Controller struct {
@@ -34,6 +34,7 @@ func NewController(config ControllerConfig) (*Controller, error) {
 	{
 		c := controllerresource.Config{
 			BaseDomain:       config.BaseDomain,
+			Provider:         config.Provider,
 			CreatePVC:        config.CreatePVC,
 			StorageSize:      config.StorageSize,
 			K8sClient:        config.K8sClient,