+++ service/controller/resource/scrapeconfigs/resource.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"reflect"
-	"strings"
 
 	"github.com/giantswarm/k8sclient/v3/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
@@ -12,18 +11,27 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	"github.com/giantswarm/prometheus-meta-operator/pkg/templates"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/generic"
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
 )
 
 const (
 	Name         = "scrapeconfigs"
-	templatePath = "/opt/prometheus-meta-operator/files/templates/additional-scrape-configs.template"
+	templatePath = "/opt/prometheus-meta-operator/files/templates/additional-scrape-configs.template.yaml"
 )
 
 type Config struct {
 	K8sClient k8sclient.Interface
 	Logger    micrologger.Logger
+	Provider  string
+}
+
+type TemplateData struct {
+	APIServerURL string
+	Provider     string
+	ClusterID    string
+	SecretName   string
 }
 
 func New(config Config) (*generic.Resource, error) {
@@ -33,10 +41,12 @@ func New(config Config) (*generic.Resource, error) {
 	}
 
 	c := generic.Config{
-		ClientFunc:     clientFunc,
-		Logger:         config.Logger,
-		Name:           Name,
-		ToCR:           toSecret,
+		ClientFunc: clientFunc,
+		Logger:     config.Logger,
+		Name:       Name,
+		ToCR: func(v interface{}) (metav1.Object, error) {
+			return toSecret(v, config.Provider)
+		},
 		HasChangedFunc: hasChanged,
 	}
 	r, err := generic.New(c)
@@ -47,13 +57,14 @@ func New(config Config) (*generic.Resource, error) {
 	return r, nil
 }
 
-func toSecret(v interface{}) (metav1.Object, error) {
+func toSecret(v interface{}, provider string) (*corev1.Secret, error) {
 	cluster, err := key.ToCluster(v)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
-	scrapeconfigs, err := parseTemplate(cluster.GetName())
+	var clusterID = cluster.GetName()
+	scrapeConfigs, err := renderTemplate(clusterID, provider)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
@@ -64,7 +75,7 @@ func toSecret(v interface{}) (metav1.Object, error) {
 			Namespace: key.Namespace(cluster),
 		},
 		Data: map[string][]byte{
-			key.PrometheusAdditionalScrapeConfigsName(): []byte(scrapeconfigs),
+			key.PrometheusAdditionalScrapeConfigsName(): []byte(scrapeConfigs),
 		},
 		Type: "Opaque",
 	}
@@ -72,24 +83,31 @@ func toSecret(v interface{}) (metav1.Object, error) {
 	return scrapeConfigsSecret, nil
 }
 
+func renderTemplate(clusterID string, provider string) (string, error) {
+	var templateData = TemplateData{
+		APIServerURL: fmt.Sprintf("master.%s", clusterID),
+		ClusterID:    clusterID,
+		Provider:     provider,
+		SecretName:   key.Secret(),
+	}
+
+	content, err := ioutil.ReadFile(templatePath)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+
+	template := string(content)
+
+	scrapeConfigs, err := templates.Render(template, templateData)
+	if err != nil {
+		return "", microerror.Mask(err)
+	}
+	return scrapeConfigs, nil
+}
+
 func hasChanged(current, desired metav1.Object) bool {
 	c := current.(*corev1.Secret)
 	d := desired.(*corev1.Secret)
 
 	return !reflect.DeepEqual(c.Data, d.Data)
-}
-
-func parseTemplate(clusterName string) (string, error) {
-	template, err := ioutil.ReadFile(templatePath)
-	if err != nil {
-		return "", microerror.Mask(err)
-	}
-
-	var scrapeconfigs string = string(template)
-	var apiServerUrl string = fmt.Sprintf("https://master.%s", clusterName)
-
-	scrapeconfigs = strings.ReplaceAll(scrapeconfigs, "<SECRET_NAME>", key.Secret())
-	scrapeconfigs = strings.ReplaceAll(scrapeconfigs, "<API_SERVER_URL>", apiServerUrl)
-
-	return scrapeconfigs, nil
 }