+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Scraping of legacy docker, calico-node, cluster-autoscaler, aws-node and cadvisor (#78)
+
 ### Changed
 
 - Moved prometheus storage from `emptyDir` to a `persistentVolumeClaim`